[INTERNAL] Demo Kit: Build card removed from Tools page
There is a new a new requirement, which states that the "Build" card
should be removed from the Tools page of the Demo Kit.

This ruins the whole look and feel of the layout in the Tools page.

This is why the following changes to the layout are made:

1. Large breakpoint:
-Previously: 1row - 2columns, 2row - 3columns, 3row - 2columns
-Now: 1row - 3columns, 2row - 2columns, 3row - 3columns

2. Extra Large breakpoint:
-Previously: 3rows with 3columns each
-Now: 2rows with 4columns each

3. Cards order:
-Previously: Web IDE, Theme designer, UI5 Tooling, UI5 Inspector, Icon
Explorer, Card Explorer, Theme Parameter Toolbox, Support Assistant

-Now: Web IDE, Theme designer, UI5 Tooling, UI5 Inspector, Card
Explorer, Support Assistant, Icon Explorer, Theme Parameter Toolbox

All of the above are synced and approved by Visual Designer.

Change-Id: I55210dd53a1bbf3b33169eb6b8905a74223ca2d3
JIRA: BGSOFUIPIRIN-4514