[INTERNAL][FIX] sap.f.ShellBar: Fixed truncating of title
Problem:
Title is truncated when there is available space.

Root cause:
Title has fixed width which is calculated wrong. Font is not loaded and it
is still not applied to the text which results to wrong value.

Fix:
Now, the title width is cought when the theme is applied.

BCP: 2070095624
Change-Id: I12efc951080ed7222306a591ddd6b367a60ce9e8