[INTERNAL] m.Table: corrections for the autPopinMode
The autoPopinMode relied whether the table items are loaded. But in the
SmartTable and the MDCTable, it is possible to configure the table with
autoBinding disabled. Hence, the autoPopinMode of the Table is modified
to not rely on the table items. To optimize the calculations, a function
is introduce to only recalculate incase the column aggregation changes.
JIRA: CPOUIFTEAMB-1469

Change-Id: Ic4b89c06da2fc207bfceaebd3fe8ecf212a28b71