[INTERNAL] sap.m: explicitly load sap.ui.unified before dependencies
sap.m does not declare a dependency to sap.ui.unified, but some modules
in sap.m have dependencies to modules in sap.ui.unified. This causes
individual requests when sap.ui.unified has not been loaded in time.
To avoid slow cascades of requests, sap.ui.unified is explicitly loaded
synchronously before the dependent modules are resolved.

Loading the library synchronously is only a last resort, the correct
solution is to declare a dependency to sap.ui.unified on consumer side
(e.g. in the manifest of an app).

Change-Id: I8288a231c78e1d02a05273071809eb84638066f4
BCP: 2070110240