[FEATURE] sap.ui.integration.widgets.Card: Handle Action.Submit
At this point, AdaptiveCards provide Submit action which just provides an
object/map with field->value pairs of the input fields.
Host environment could benefit from that action by providing a Host config
which handles the data submission.

When it comes to the ones that provide the manifest and define the card,
there's no way to control and define how this Submit action would be handled or
configured.

This change introduces a manifest configuration which would allow manifest
authors to define a destination with a basic configuration for the Submit
action.

JIRA: BGSOFUIRILA-2749

Change-Id: I9ae60c2ffe3d0fce6af5015031d796ecdddc32d5