[INTERNAL][FEATURE] sap.ui.rta: enabling validators for rename in RTA
control owners can now add a validator (or multiple) in the designtime
metadata for the rename action in key user adaptation. This can be
either a custom validator or a predefined validator.
The predefined validator added in this commit has the name "noEmptyText"
and simply checks the new text is not empty

Change-Id: Iaae653fd454f54f7fc680a1ea6978ef1ddb56ef2
JIRA: UICSFLEX-6890