[INTERNAL] sap.ui.integration: improvement for escapeParameter utility
Added a validator function to give the consumer of the utiliry whether
ability to have custom logic which parameters should be escaped. E.g.
in case of PropertyEditor wrapper we do escape not only by name (as it
was before), but also checking where "type" property of the specified
configuration object is registered in the PropertyEditorFactory.

Additionally, escape utility now iterates through all the plain object
which are specified in mSettings for the ManagedObject and make an
escape attempt based on the validator decision.

Change-Id: Ib96d94d4a752f0c654ddac43d453528efaeea8d8
JIRA: UICSFLEX-6185