[INTERNAL][FIX] sap.m.Select: Fixed scrolling to selected item
Problem:
Select wasn't scrolling to its selected item if it is not in the
visible content of the Popover.

Root cause:
In previous versions, vertical scrolling of Popover was turned
off automatically, but after changes in the behavior
it had to be changed explicitly.

Fix:
Now, vertical scrolling of the Popover is disabled.

BCP: 2070058798
Change-Id: Ib866f6dddc3614e0a45f7534f9ecda8ebd547751