[INTERNAL][FIX] sap.uxap.AnchorBar: Refactored AnchorBarButton styles
Problem: On hover and active states on some of the elements inside the
AnchorBarButton there is unwanted background color. The color should be
always transparent/none on all themes.

Root Cause: Regression introduced from the base styles of the
sap.m.Button.

Solution: We set the "background", "border" and "text-shadow" styles
of all of the "sapMBtnInner" elements inside AnchorBarButton to "none"
with "!important". We do this in the base less file of the
AnchorBarButton. This way we don't need to adjust the above styles
every time the sap.m.Button decides to implement stronger selector.
Additionally redundant(now) code which ensures the same is deleted from
all of the supported themes.

Change-Id: Ied34dd71f7209f1be2ad21fb3fbafc3f942be4e6
BCP: 2080116759