[FIX] sap.m.OverflowToolbar: Empty Popover is fixed
Problem: When there is not enough space for a control with
OverflowToolbarPriority=Disappear, it disappears and an overflow button
appears. When the button is clicked, the opened Popover is empty.

Solution: Before showing the overflow button, we check if there is at
least one control that can be shown in the Popover (controls with
OverflowToolbarPriority=Disappear are never shown in the Popover).

Change-Id: If331cb194f4f260bd3e4e392f84d38a162ca4fd3
BCP: 2070107682
(cherry picked from commit 73c5aad45dccc345f90aad6b66db63680b93fa97)
CR-Id: 002075125800002697052022
BCP: 2270083289