[INTERNAL] Table: QUnit tests - Ensure that libraries are loaded upfront
Avoid usage of other libraries in test whenever possible. Otherwise, make sure
that all required libraries are loaded upfront.

If libraries are not loaded before the test starts, the library theme will be
loaded in parallel to the test execution. Once theme is loaded, the table will
react to this (onThemeChanged), invalidate and rerender. If this happens to a
rendered table that is tested in a test which does not expect rerendering, the
test will most likely fail.

Change-Id: If046eb926ef4730750fbdb7dae5af87c0128c305