[INTERNAL] m.Table: minor corrections to the autoPopinMode
When the autoPopinMode is toggled after the control has been rendered, e.g,
initially true, then false and true again, the recalculation is not applied
as expected. This change is related to the SmartTable where the column's
demandPopin and minScreenWidth settings are influenced to deactivate the
popin behavior. However, stand-alone Responsive table does not support the
toggling of the autoPopinMode. The same is also documented in the autoPopinMode
API docu.

Change-Id: If3dd9478fc7b35442576ca3d43dba7591af56a15