[INTERNAL] m.WheelSlider: reference imported renderer in metadata
The WheelSliderRenderer does not expose itself under a global name
(which is good). But due to this, the RenderManager does not find it and
assumes it has to be loaded synchronously, although the WheelSlider
class has imported it already.

Fixed by referencing the renderer in the metadata of the control.

Change-Id: I1304a2e5847beaad54d1097da0c77436212b2edd