[INTERNAL] sap.ui.core: use a non-bundle URL for dummy requests
Performance trace tests trigger dummy XHRs to check that request are
properly monitored and attributed (e.g. with FESR or passport header).
To ensure that the request are sent out, they use the URL of a wellknown
resource (sap-ui-core.js) and add a varying cache-busting token.

But in test environments where bundles like sap-ui-core.js are created
on-the-fly by the server, this can result in many creations of the same
bundle. Therefore, the tests now use 'resources/ui5loader.js', which is
not a bundle.

Change-Id: Ie70821319f575e0aca60ebcb99db0ce18dc9c6d2