[FIX] sap.m.OverflowToolbar: Cutting of action buttons is fixed
Problem: There are some Samples in the Demo Kit with
FlexibleColumnLayout, which has a DynamicPage in its middle column. When
the user opens the Sample, the Sample.controller of the Demo Kit fires a
theme change event. The OverflowToolbar, placed in the DynamicPageTitle,
listens for this event. Upon theme changing, the OverflowToolbar resets
itself and calculates the size of its overflow button. However, a timing
issue sometimes appears. When the calculation of the overflow button is
done, the middle column of the FlexbileColumnLayout is still with
display=block, but visibility=hidden. As a result - the overflow button
does not have width, but is still has margin (because of its styling).
And this margin is wrongly cached as total width of the button. Later
on, this wrong value leads to visual degradation and cutting of some of
the actions buttons of the DynamicPageTitle.

Solution: Now, when calculating the width of the overflow button, we
consider only the case when the button has its own width. If it has only
margin, we do not cache the width value.

Change-Id: Ic6683277d9c1ca942d6b54a8a9ed21d7b103ae11
BCP: 2070009883