[INTERNAL] Demo Kit: Deprecated controls are now marked in the samples
If a control is deprecated, now there is a sap.m.MessageStrip with
"Error" state and "Deprecated" text, which shows the user that
the selected control is deprecated when its samples are opened.

The sap.m.MessageStrip is positioned above the table with the samples
themselves in the "Samples" section.

Change-Id: I9392eb00c2cc942befa81839ac99a3cc716b43af
JIRA: BGSOFUIPIRIN-4527