[FEATURE] sap.m.BadgeEnabler: New Badge API introduced
The new badge concept contains of a functionality to show and hide badge element of certain
controls, responsible for indicating an event on the control.
It requires the control to implement sap.m.IBadge interface and to "activate" sap.m.BadgeEnabler
helper with specific configuration, depending on UX requirements and control DOM structure.
Also the control should have customData aggregation attached with object of type
sap.m.BadgeCustomData aggregated, which receives only 'value' property, holding information
shown in the badge.

Change-Id: I1ba6ccce47d7e00d4833ea82b95466e9755fe733