[INTERNAL] sap.f.FlexibleColumnLayout: Focus handling improvement
Problem: When navigating back to a column and there is not
previously stored element to focus, focus should be placed on the
first focusable element inside the current page of the column.

Solution: Focus is now placed on the correct element when
restoreFocusOnBackNavigation is true but there is no stored element
from previous navigation.

Change-Id: I4050c3f75799e44348b05569c5fd7ff88491078d
JIRA: BGSOFUIPIRIN-1485