[INTERNAL] ListBase: render dummy keyboard handling area
Regardless whether items or noDataText is rendered or not, the dummy
keyboard handling area must be rendered to ensure that the focus handling
works as expected. In case the dummy area of the List is not rendered and
backward navigtion is perfromed on the first item of the list, then the
focus is set on the <ul> element which leads to unwanted accessibility
announcement. The rendering of the dummy area in the list take care of
correctly forwarding the foucs to the relevant control and prevents unwanted
accessibility announcements.
BCP: 2070064044

Change-Id: I631499ab50858a9e00cb2ffc0fa5718609660389