[FIX] model.Model: avoid too many setTimeout/clearTimeout calls
With the old implementation, consecutive calls to addBinding /
removeBinding could cause many setTimeout / clearTimeout calls where
each _cleanUpBindings execution had to cleanup only very few bindings
(most of the time only 1). This happened especially in the context of
XML templating where many OneTime-bindings are created.

As setTimeout / clearTimeout seems to be expensive in V8 when used
massively, the strategy has been changed:

If there's a pending setTimeout when _cleanUpBindings is called, the
timeout is not cleared, but kept for reuse if another removeBinding
follows.

Additionally, to-be-removed bindings are stored in a Set, not an array.
This makes the loop in _cleanUpBindings less expensive if many bindings
have to be removed.

Change-Id: Iaeb6fc16ca566bca7043dc4bb4f1797e96f10cbb
BCP: 2070128920
BCP: 2070128967