[INTERNAL] sap.ui.integration.widgets.AdaptiveContent: Move enableMarkdown settings at sap.card level
We're not able to extract the enableMardown property out of
the configuration as it represents manifest's content part.
The AdaptiveContent is still not an aggregation of the Card.
We should extract the configuration straight from the Card,
but when it's plready a parent of the Adaptive Content.
This is guaranteed in the onProcessMarkdown hook.

Change-Id: I8377ce3c9507785e28976d2e2d6b5de5ebbd646d