[INTERNAL] Demo Kit: Performance optimization of API Reference page
Problem: Every time a new API Reference page of a Control is opened, the
XML view is getting destroyed and a new one is created. This takes too
much time and slows down the page.

Solution: The XML view is created only once (the first time an API
Reference page of any control is opened). After that, only data, which
is binded, is getting changed.

Change-Id: I84b5013b4fafccad4285eeddb708b10e47daa3d9
JIRA: BGSOFUIPIRIN-2406