[INTERNAL] sap.ui.integration.widgets.Card: Fix failing test CardLoading
Problem: When the project is built and library-preloads are used
plus browser cache, the test failed with error:
"assert.async called too many times".
This is because onAfterRendering is happening multiple times before the
test finishes.

Solution: In order to test the Loading, it is enough to check only
the first rendering, and unsubscribe for all subsequent.

Change-Id: I1927d7936955cb351e850eef014cec7d3d595762
BCP: 2070134915