[FIX] sap.f.DynamicPageTitle: GenericTags overflow correctly now
Problem: When GenericTags are placed in an OverflowToolbar (part of the
DynamicPageTitle's content aggregation), once they go into the
OverflowToolbar's Popover, they never get out of it, even if there is
enough space.

Solution: The problem happens mainly due to three reasons:
- The GenericTag's wrapper is an inline-flex element with 100% width,
but it's parent does not have specified width. This means that the
wrapper is trying to take the whole width of its parent, but the
parent's size depend on its own content, so none of the elements "knows"
how wide it should be. Now the GenericTag's wrapper has display=flex to
solve this problem.
- There is a previous fix which invalidates the DynamicPageTitle, when
there is a change in the size of the content of the OverflowToolbar and
this content contains GenericTags. This fix was introduced for
situations when GenericTags are getting rendered with no data at first
(so after the _controlWidthChanged event is dispatch from the
OverflowToolbar, the content part of the DynamicPageTitle has flex-basis
set to the desired width). However, when the data comes (with a delay),
a problem with space distribution appears as the GenericTag cannot
determine its own width (described in the bullet above), when it's
placed in the flex area of the DynamicPageTitle.
- There is a new right padding on the OverflowToolbar, placed in the
content aggregation of the DynamicPageTitle, which comes from a Design
specifications. However, when OverlowToolbars in content and action
parts of the DynamicPageTitle have paddings, the space they need is not
calculated correctly, so the _controlWidthChanged event is fired to the
DynamicPageTitle with wrong value. As a result, the DynamicPageTitle
sets wrong flex-basis, which results in either all GenericTags go in the
OverflowToolbar's Popover, either once they go in, they never go out.
It seems this right padding is needed for the OverflowToolbar in the
content part because the GenericTag has box-shadow, instead of borders.
In this case, its box-shadow gets cut, as the OverflowToolbar is wide
exactly as much as its content is wide. So the box-shadow disappears.
In order to keep the box-shadow, the right padding is kept and a new
left padding is also added. Now when the OverflowToolbar dispatches its
_controlWidthChanged event it includes its left/right paddings.

JIRA: 2070141690
Change-Id: I37dfccc831823edc16161e77212dd0bc530af9e0