[FIX] TreeTable: Fix single selection if collapseRecursive is false
The selection API of the TreeBinding only affects the currently "reachable"
part. Therefore, nodes that are the child of a collapsed node cannot be selected
or deselected.

If collapseRecursive is false, the TreeBinding remembers the selection of
children of a collapsed node. When this node is expanded again, it's children
that were selected are selected again, regardless which APIs have been called in
the meantime.

If the table is in selection mode "Single", there must not be more than one
selected index.

Depends on: Ifca765711b0078e0e95df9288a24907d701edb5a

BCP: 2070052448
Change-Id: I0d574035ea06558bca54b7dd7f48321b2dfe0bfe