[INTERNAL] Table: Fix scrolling with the mouse wheel on the scrollbar
The scrolling extension internally sets flags when scrolling with the mouse
wheel or keyboard. These flags should, for example, prevent handling of
subsequent scroll events of the scrollbar to avoid unnecessary computation and
accidental changes of the scroll position. As the flags are not reset when
trying to scroll with the mouse wheel on the scrollbar element, scrolling does
not work.

The flags are removed, because they are no longer required with the
process-based implementation of the scrolling extension.

Change-Id: I785068962363ba0cc09acef02ff62b5b0a5b2baa
BCP: 2070106221