[FIX] sap.m.CheckBox: Fixed incorrect display of focus on contrast themes
Problem: Left side of the focus of the sap.m.CheckBox is incorrectly
displayed. It's under the checkbox itself when we are in a Form in all
contrast themes in compact content density.
There is a css rule for this exact issue for the SimpleForm, but not
for Form/SmartForm.

Solution: We now use the sapUiForm css class for the focus rule instead
of sapUiSimpleForm.
This way we get the correct behavior for the Form/SmartForm, without
introducing a regression for the SimpleForm, because SimpleForm also
has the sapUiForm css class.

Change-Id: I5c6128d9c16051fb1bbb73c952bbf122cc9cc416
BCP: 1980435273