[INTERNAL] sap.uxap.ObjectPageLayout: Ensure subSection fits container
Background:
Whenever the (sub)sections visibility/titles are changed, the
_adjustLayoutAndUxRules is called (to adjust the layout,
rebuild the anchorBar and update layout-dependent properties)
without rerendering the entire page.

Problem:
The value of this._bAllContentFitsContainer was not updated after change of layout
and its value was wrong.

Solution:
Update the value this._bAllContentFitsContainer in _adjustLayoutAndUxRules, after
having obtained info of the current visible (sub)section and currently selected tab.

BCP: 2070022388
Change-Id: I542dbc45a7c602801ea04df4610791f4ec8e5343