[FIX] sap.uxap.ObjectPageLayout: Early focus trigger in IE fixed
Problem: In sap.uxap.ObjectPageLayoutABHelper we have a logic, targetting
applications with "full" and "basic" animation modes, as in the first case
we are adding a delay before we focus on a section, on selecting a section,
giving the animation time to end. In the second case we simply need to trigger the
focus event immediately, which on IE gives wrong calculated value for the scrollTop
of the given section, after it is focused (8px), and the section change never
occurs.

Solution: Ensured that layout is correctly calculated even when sections
are positioned (e.g. with position:relative when a section in focused on IE).
Positions of the scrollContainer elements are now always measured with respect to the
scrollContainer, even when the scrollContainer is not their immediate offsetParent.

Change-Id: Ieee20bb2b8aa3658ea208b5d6d73d793a35ad92a
BCP: 2070134081