[FIX] DragAndDrop: The text of input fields should be selectable in FF and Edge
If an input ot textarea element is inside a draggable element, the cursor cannot
be placed inside the element, and the text cannot be selected. In some browsers,
this is a long-known bug for which a solution is not yet foreseeable.
A workaround has already been implemented for IE:
I2db5d99ec64eccc0e342017fab3da49351c30855
The workaround is now also activated for Firefox and Edge (EdgeHTML). As of
now, the issue does not occur in browsers with the Blink browser engine.

Change-Id: I62205ced308bf4920bc5adc450cb14cd56c3cdc1
BCP: 2070157420