[FIX] AlignedFlowLayout: Orchestration of the initial rendering fixed
The problem:
Incorrect width and alignment of the last item in certain complex
scenarios, for example, when the container of the AlignedFlowLayout
control is invalidated.

The root cause:
The reflow algorithm of the AlignedFlowLayout control is triggered
before the browser layout is finished.
Usually on after rendering, the browser layout is finished, but
this is not always the case.
Thus causing the reflow algorithm of the AlignedFlowLayout to
misbehave base on the actual position of the items.

The solution:
On after rendering, request the next animation frame to wait until
the browser style calculations and layout is done, then trigger
the reflow algorithm of the AlignedFlowLayout control.

Change-Id: I489361a7945b7f2d456412caae45d4b3fbd39274