[FIX] XMLTemplateProcessor: Handle Top-Level ExtensionPoints and Flexibility
Parsing of XMLViews was broken in case the XMLView contains ExtensionPoints
as top-level children.

With this change we collect all ExtensionPoints on the top-level as placeholders
and trigger the processing of Flexibility changes according to the agreed API.

ExtensionPoints inside nested UI5 control structures have not been affected by
this bug.

Change-Id: Ib5cf7bc0c37a76a95aeadb0d30477a483f775076
BCP: 2070136077