[INTERNAL][FIX] TreeTable: Avoid errors if bound to a V1 OData model
If the table is bound to a V1 OData model, the
sap.ui.model.TreeBindingCompatibilityAdapter is applied to the TreeBinding.
In this case, errors occur when updating the row state and destroying the
BindingSelection plugin.

Side note: Even though this change solves the issue in the linked incident, it
is not actually the solution! The app overrides TreeTable#getBinding and applies
the sap.ui.model.TreeBindingCompatibilityAdapter instead of the
sap.ui.model.ClientTreeBindingAdapter to the client binding. Therefore, the
issue also occurs with a JSONModel - in this app.

Change-Id: I3f67d993afd2fd06c92d3aaee481bdb3b9c2ab7b
BCP: 2070171260