[INTERNAL] sap.ui.core.sample.odata.v4.SalesOrders.tests.WriteNonDeferredGroup: make OPA more robust
When the initially shown sales order list already has some bound messages,
then the OPA failed because the message details for the triggered RAISE_ERROR
message are not shown automatically but have to be selected first.
This is because the message popover automatically displays the
details if there is exactly one message in the message popover.

Test:
http://localhost:8080/testsuite/test-resources/sap/ui/core/demokit/sample/odata/v4/SalesOrders/Opa.qunit.html?filter=Write&realOData=true

PS1: first attempt
PS2: reason clarified
PS3: review comments TCH

Change-Id: I8d5515320fcddccfaf9168020ab5fd25bc0f1356