[INTERNAL] Table: Remove active state CSS from the "SelectAll" checkbox
The SelectAll checkbox does not need an active state, it's a checkbox.
In case of the MultiSelectionPlugin's "deselectAll" button, the plugin's CSS
takes care of the active state for this button.

The CSS for the active state of column header cells is moved to the relevant
class to avoid potential side effects on other header cells in the future.

Change-Id: I5a6084beffd6f7f65076ac36d8afefa08eaf7ef1