[INTERNAL] sap.ui.unified.calendar.Month: unit tests now run successfully on IE
Problem description:
A test was counting on that particular DOM element would exist after
a week number is pressed. The control may get rerendered when the
first week number is pressed, in which case this DOM node would not
exist any longer.

Solution:
The test now uses the second week number in order to make sure that
rerendering would not appear.

BCP: 2070151438

Change-Id: I45998c3d71e29a7a6594219a6658f1965c1fc30e