[INTERNAL] sap.ui.fl: make DependencyHandler more resilient
this should be reverted as soon as we get rid of FakePromise in favor of
real Promises.

In general it is important that per control the change applying is not
disturbed by a different control that is being created. When
ChangeHandler create controls synchronously, and those controls have
changes on them, this is not guaranteed anymore, as long as we are
working with FakePromises.

Change-Id: I15a361fd5660de52f33f5dd9d03133a207501728
BCP: 2070126206