[FIX] sap.m.FactFilter: selected items text in the facet buttons is now correct
Problem description:
When sap.m.FacetFilterList default filtering behavior is prevented inside the
"search" event handler function, then all list items are not loaded initially.
For this reason we cant't know how many items there are in the data model,
and the logic, which checks if all items are selected is not working correcty.

Solution:
List items are binded and filtered with a custom static or dynamic filter
inside the "search" event handler function. We can only get the list items count
from the list items binding after the filtering is performed.
So in this case all items are those items which are currently loaded
and not all items from the data model.

Change-Id: I0415d6568a9bf1e882c4ab654691f4ff46366c30
BCP: 2070160117