[FIX] sap.f.DynamicPageTitle: Qunit test no longer fails at IE11
Problem: There is a known issue with shrinking priorities and flex containers, which
are also flex children. Its root is that the default min-width of the element is
set depending on its content. We overwrite and fix that using "min-width: 0" rule,
to force the element to shrink, regarding his shrink-factor in a situation, when
 there is not enough space to fit its content. Anyway, IE11 doesn`t read it
properly.

Solution: We use the same fix, only using "%" unit for the IE to make calculation
based on it.
https://github.com/philipwalton/flexbugs#1-minimum-content-sizing-of-flex-items-not-honored

BCP: 2070187869
Change-Id: I69c1b77885d87c0132ff556120c13f5fd17591e4