[INTERNAL] sap.uxap.ObjectPageLayout: Removed redundant layout adaptation
Background:
In order to obtain the correct snap/expand breakpoint, the ObjectPage
measures in advance the height of the header in the alternative (spanned vs. expanded)
state. To do this, it creates a clone of the headerTitle, puts it in teh alternative
(spanned vs. expanded) state and measures its height

Problem:
To adapt the layout of the clone of the headerTitle, the
_adaptLayoutForDomElement is called on the cloned headerTitle.
The function also adapts the headerTitle actions, which is
unnecessary, as the visibility of the actions has no effect
on the final height.

Solution:
Replaced _adaptLayoutForDomElement with only the function that is
needed to obtain the expected height. The remaining operations in
_adaptLayoutForDomElement (that have no effect on the final height)
are thus skipped.

Change-Id: I38a8fcbdf2e85611002fb0385704cb1e60c89953