[FIX] sap.uxap.ObjectPageLayout: Fix lazy loading in IconTab mode
Problem: When choosing sections from the IconTabBar, they are
not loaded.

Solution: _grepCurrentTabSectionBases method of the ObjectPageLayout returns
only the SubSection, but not the Sections. Now it is adjusted to return
also the Sections as the getSubsectionsToPreload method of the
LazyLoading expects also to receive Sections.

Change-Id: I88d8e87b924bc730fedc1bb3df9b20fcf4ad02d8
BCP: 2070191938
(cherry picked from commit 40b213c5908844996ccce1ad8dbf050f3875a55f)
CR-Id: 002075125800002753112020