[INTERNAL] Demo Kit: Improved event info in the transformation script
Previously: We were checking if parameters exist in the Symbol, but
we weren't checking if the exact param (oParamInSymbol[sParam]) actually
exists.

Now: We make the "oParamInSymbol" to be the exact param, and not the
object which contains all the params of the event. This way we avoid
a pitfall where we try to access an undefined property of the old
"oParamInSymbol" object.

Change-Id: I70520f256c1e28a61a9eb0e67c6bb082755732dc