[INTERNAL] sap.uxap.AnchorBar: Add missing text in messagebundle.properties
Background: After refactor of AnchorBar roles a translation was placed in the
wrong file and it worked only for 'en' langauge.

Problem: A text was not translated properly in most of the languages.

Solution: Place the translation in messagebundle.properties file.

Impact: The regression decreases the accessibility of the control, by
not reading the correct text when screen reader is on.
This may bring confusion in end users and loss of navigation flow.

Risk: The modified texts do not affect functionality and the risk should be very low in regards to translations.

Change-Id: I3d26b7c8cbcd6c1348624a7a2ab2ed00f1523fd9
CR-Id: 002075125900003912622020
BCP: 2080310286
JIRA: BGSOFUIPIRIN-4515
(cherry picked from commit 54c37709b3e8611fabb43c4d5b586ed27f08c334)