[FIX] sap.uxap.ObjectPageLayout: Revert no lazyLoading during animation
This reverts commit 14138755c596d5d956081dd5d429b66cb3920f7f.

Reason for revert: The change uses criterion for end of the animated scroll that is not reliable
(as it expects the animation to end after the requested animation duration, but the animation may
take longer, esp. on a loaded processor).
So, a separate change is proposed to allow attach callback for the exact end of the scroll
animation. The change is under discussion as the changed control
(sap.ui.core.delegate.ScrollEnablement) is not owned by our team.
Temporarily revering this change until a final solution is found.

Change-Id: Ic0bd7bfe04a5f5219571c1b1115331aead59b751