[FEATURE] sap.m.IconTabBar: introduce property tabNestingViaInteraction
This property restricts by default the end-user's ability to hierarchically nest
IconTabFilters within each other by using Drag and Drop.
This behaviour becomes opt-in for application developers who wish their
users to be able to nest their tabs by drag and drop.

The reason for introducing this as an opt-in behaviour, is that it now
allows in all instances of IconTabBars to now be able to nest tabs out
of the box (when the property enableTabReordering is true), of which
applications have not been designed with this ability in mind.

Also fixed an issue with dragging an item over a tab which has children but has
no content, such as the Overflow tab.

JIRA: BGSOFUIRODOPI-2338
BCP: 2070219677
Change-Id: I789696dc3edf76092d19e5cbeb598ebcaa6d3678
CR-Id: 002075125800002398432020

(cherry picked from commit 7f70affd8b2760662d9763f184a98935c874412d)

Change-Id: I033a2b37018f57c1c89b65a52d4db5719dc560ec