[INTERNAL] Demo Kit: Documentation tab search improved
Problem: When the client search for something, there is a probability
that on the way to the search query, there are some branches which
end up empty, even though they are populated if you refresh the
page after selecting them.

Root Cause: The missing leafs of those branches are lost during
filtering.

Solution: In order to keep the information of the missing leafs, when
building the search criteria (name), we take into account the whole
path to the item, not just the parent.

Change-Id: Ib269da27dccaf7557aec62f5a334031cc8f20dd3
BCP: 2070199925