[INTERNAL] Table: Adjust rows update to renderer API version 2
If properties of cell controls are changed, they are re-rendered asynchronously.
After changing to the new rendering mechanism, this is now also the case for
controls that are officially supported in the table.
When updating the rows, we have to wait for this rendering before things like
synchronization of row heights can be done.

Change-Id: I415cb26fa1234b88aa6dbf681001a6f5563cf677