[FIX] sap.m.App: QUnit failure fixed
Problem: QUnit is failing when executed in a separate Browser window.

Solution: Failing assertion is removed as it has never been supposed to
pass when the test is executed in a separate Browser window. It seems
the orientation change event is always fired when opening a new Browser
window. Moreover, this QUnit is testing depricated functionality, which
is supposed to not be used at all.

Change-Id: If742266dae9f966f2fa229bdddcc67fc74fb40f2
BCP: 2070153495
CR-Id: 002075125800002318102020