[INTERNAL] ListBase: fix range selection
When the user selects an item in the list and switches the browser tab via
keyboard shortcut CTRL + SHIFT + TAB, this action created the Range selection object.
Due to the range selection object, the user is not able to deselect the item
when coming back the browser tab where the UI5 app is open.
Hence to fix the issue, the rangeSelection should be prevented when the
SHIFT key is used with any addition key combination.
BCP: 2070186307

Change-Id: Ie48c6c51da8eb7879ea4985c906586336f7e7911