[FIX] sap.uxap.ObjectPageLayout: Corrected error upon rerendering
Problem:
Upon internal rerendring of the page, the app may request to scroll to a section
that requires snapping of the header. If the app made that call before the
'onAfterRendring' hook of ObjectPage => error because the DOM elements
cache is not updated yet.

Solution:
Check for existence of the cached DOM element before manipulating
the DOM element.

Change-Id: Ibfc77a26a9f37846f22599acbda872234ce2b19e
BCP: 2070205321