[INTERNAL] ListBase: minor rangeSelection fix
A minor fix for the rangeSelection. RangeSelection should also be prevented
when the user has selected an item and performs fast navigation with SHIFT + F6.
The above action also prevented the user from later deselecting the selected
item. This is now fixed with this change. Along with this, the qunits are
refactored and controls are placed in qunit-fixture DOM element.

Change-Id: I4056f1020eb275bcb2fd5e2f85ea8e7370ceecb1