[INTERNAL] Prefer standard way of calling super methods
If possible, a class should call inherited method implementations always
on the direct super class. This makes the call not only independent from
changes in the ancestor classes, but also makes it clear that the call
is compatible with the 'super' keyword in ES6 class syntax.

For the same reason, the following pattern for overriding an inherited
method should be avoided:

  MyClass.prototype.savedMethodA = MyClass.prototype.methodA;

  MyClass.prototype.methodA = function() {
    this.savedMethodA(); // call inherited
  };

It doesn't translate well into ES6 class syntax. Instead of keeping a
local reference to the inherited method, it should be called on the
direct prototype of the direct superclass.

Only if the nearest implementation of a method in the inheritance
hierarchy should be skipped (because it doesn't fit the needs of the
current class), then calling the method on a class other than the direct
base class makes sense. In that case, the reason for the deviation from
the standard pattern should be documented.

BTW: Sticking to the standard pattern can help to reduce the necessary
dependencies of a class.

Change-Id: I7a7712906bfebaeaedca0bdf217cb2ac884de444