[INTERNAL] all: Fix library info in runtime metadata
When a ManagedObject class resides in a library sub package with a lower
case name, then the 'library' property must be specified in the metadata
of the class.

Otherwise, the framework will guess a wrong library name and, if the
class is a control or element, assign the class to that non-existing
library. Support rules might detect this and mistakenly report it as an
'unused library'.

Strictly spoken, the library property is only needed for controls and
elements as other ManagedObjects are not assigned to libraries. But as
the calculated library name is also wrong for ManagedObjects, this
change fixes them, too.

Change-Id: Ie884493be6d400a19490d29d81c43c591a56aba5
BCP: 2080285954