[INTERNAL] Replace usages of internal function Metadata.createClass
Metadata.createClass is a private, internal function of the sap.ui.base
package and must not be used in control or application code. instead,
sap.ui.base.Object.extend should be used.

All such existing usages have been migrated with this change. For best
compatibility, getInterface is implemented in each affected class to
hide the only functionality that is inherited from sap.ui.base.Object
and which might cause issues. The second inherited functionality, isA,
is kept visible as it might be useful anyhow.

The remaining usage in mvc/Controller is kept as it does not start a new
class hierarchy but creates a subclass of an existing controller
extension. This call intentionally bypasses the final flag for
controller extensions and therefore can't be replaced by calling extend.

Change-Id: I4ed7820f256111a30e569ea5137d0329d3c8195d