[INTERNAL] DataStateIndicator: hidden properties messageLinkText and messageLinkVisible added
Private properties messageLinkText and messageLinkVisible are added to
the DataStateIndicator plugin.
The DataStateIndicator plugin creates a sap.m.Link control when the messageLinkText
property is set. This link control is added to the "link" aggregation of the
MessageStrip control. The visibility of the link control can be controlled
by the messageLinkVisible property of the DataStateIndicator plugin.
JIRA: CPOUIFTEAMB-1530

Change-Id: I3629c952c37570980ec755604f754cc75b304f59