[INTERNAL] sap.ui.rta Fix bindingPath in Reveal
Reattaching bindingPath to reveal items in add dialog as they are used
by Opa5 tests and help debugging. However only the bindingPath of the
element that is enhancing a revealable item is attached. Further
the control attaching logic is prefixing the element better to avoid
name collisions with control properties.

Change-Id: I6c49f18dc163d84030ecd704ff5be6c03e3306ed
BCP: 2070227277