[INTERNAL][FIX] sap.ui.integration: Fix wrapper padding for touch devices
When a toolbar and form elements are set on a form, they are wrapped
in a panel.
Panels which are the first child of a toolbar receive an additional
negative padding on touch devices, which leads to labels being cut off
inside PropertyEditors wrappers in the card editor.
This can be avoided by setting the FormContainer title directly.

Change-Id: Ife9305c9141bc6f0041ff133c187112ac7a95d04
BCP: 2070215634